+++ pybamm/parameters/bpx.py
@@ -74,10 +74,6 @@ def _bpx_to_param_dict(bpx: BPX) -> dict:
     # i.e. a default C-rate of 1
     pybamm_dict["Current function [A]"] = pybamm_dict["Nominal cell capacity [A.h]"]
 
-    # number of electrons in reaction (1 for li-ion)
-    for domain in [negative_electrode, positive_electrode]:
-        pybamm_dict[domain.pre_name + "electrons in reaction"] = 1.0
-
     # activity
     pybamm_dict["Thermodynamic factor"] = 1.0