+++ tests/unit/test_solvers/test_solution.py
@@ -188,8 +188,8 @@ class TestSolution(unittest.TestCase):
         np.testing.assert_array_equal(sol.cycles[0].y, sol.y[:, :len_cycle_1])
 
         self.assertIsInstance(sol.cycles[1], pybamm.Solution)
-        np.testing.assert_array_equal(sol.cycles[1].t, sol.t[len_cycle_1 - 1 :])
-        np.testing.assert_allclose(sol.cycles[1].y, sol.y[:, len_cycle_1 - 1 :])
+        np.testing.assert_array_equal(sol.cycles[1].t, sol.t[len_cycle_1:])
+        np.testing.assert_allclose(sol.cycles[1].y, sol.y[:, len_cycle_1:])
 
     def test_total_time(self):
         sol = pybamm.Solution(np.array([0]), np.array([[1, 2]]), pybamm.BaseModel(), {})